+++ test.js
@@ -1,5 +1,5 @@
 // tooling
-const ESLit = require('.');
+const eslit = require('.');
 const readFile = require('./lib/read-file');
 
 // constants
@@ -17,14 +17,14 @@ const testData = {
 };
 
 // testing
-ESLit.import(testPath, testData).then(
+eslit(testPath, testData).then(
 	(result) => readFile('test/basic.expect.html').then(
-		(expect) => result === expect ? result : Promise.reject(`Result does not match expectation\n${ JSON.stringify({
+		(expect) => result.slice(0, -1) === expect ? result : Promise.reject(`Result does not match expectation\n${ JSON.stringify({
 			result,
 			expect
 		}, null, '  ') }`)
 	)
 ).then(
 	(result) => console.log(`${ result }\n✔ Passed`) || process.exit(0),
-	(error) => console.log(`${ error }\n✖ Failed`) || process.exit(1)
+	(error)  => console.log(`${ error }\n✖ Failed`)  || process.exit(1)
 );